refactor(desktop): move search selects into entities
