+++ apps/desktop/src/demo/UiDemoPage.tsx
@@ -48,7 +48,7 @@ import {
   ImportAssetsModal,
   ImportCompletedModal,
 } from '../features/library-workspace/import';
-import { FolderSearchSelect } from '../features/library/components';
+import { FolderSearchSelect } from '@/entities/folder';
 import { CroquisStartModal } from '../features/croquis/ui/CroquisStartModal';
 import {
   AnalyticsGraphWrapper,
