+++ apps/desktop/src/entities/folder/index.ts
@@ -1,0 +1,4 @@
+export { filterFolderSearchOptions } from './model';
+export type { FolderSearchFilter, VirtualFolder, VirtualFolderKind } from './model';
+export { FolderSearchSelect } from './ui';
+export type { FolderSearchSelectProps } from './ui';
