+++ apps/desktop/src/entities/folder/model/index.ts
@@ -1,0 +1,3 @@
+export { filterFolderSearchOptions } from './folderSearch';
+export type { FolderSearchFilter } from './folderSearch';
+export type { VirtualFolder, VirtualFolderKind } from './types';
