+++ apps/desktop/src/entities/folder/model/types.ts
@@ -1,0 +1,1 @@
+export type { VirtualFolder, VirtualFolderKind } from '@/shared/types';
