+++ apps/desktop/src/entities/folder/ui/index.ts
@@ -1,0 +1,2 @@
+export { FolderSearchSelect } from './FolderSearchSelect';
+export type { FolderSearchSelectProps } from './FolderSearchSelect';
