+++ apps/desktop/src/entities/tag/index.ts
@@ -1,0 +1,4 @@
+export { filterTagSearchOptions } from './model';
+export type { Tag, TagGroup, TagIndex, TagSearchFilter } from './model';
+export { TagSearchSelect } from './ui';
+export type { TagSearchSelectProps } from './ui';
