+++ apps/desktop/src/entities/tag/model/index.ts
@@ -1,0 +1,3 @@
+export { filterTagSearchOptions } from './tagSearch';
+export type { TagSearchFilter } from './tagSearch';
+export type { Tag, TagGroup, TagIndex } from './types';
