+++ apps/desktop/src/entities/tag/model/types.ts
@@ -1,0 +1,1 @@
+export type { Tag, TagGroup, TagIndex } from '@/shared/types';
