+++ apps/desktop/src/entities/tag/ui/index.ts
@@ -1,0 +1,2 @@
+export { TagSearchSelect } from './TagSearchSelect';
+export type { TagSearchSelectProps } from './TagSearchSelect';
